requirements.txt
langchain==1.0.3
langgraph==1.0.2
langchain-openai==1.0.1
fastapi==0.120.3
pydantic==2.12.3    
uvicorn==0.38.0
python-dotenv==1.2.1
httpx==0.28.1
minio==7.2.18
qdrant-client==1.16.0
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
motor==3.7.1